+++ app/src/main/java/com/heddrich/companion/App.kt
@@ -1,0 +1,11 @@
+package com.heddrich.companion
+
+import android.app.Application
+import com.heddrich.companion.data.CompanionDatabase
+
+class App : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        CompanionDatabase.appContext = this
+    }
+}